+++ server/src/main/java/com/smartparking/controller/SpotController.java
@@ -1,14 +1,19 @@
 package com.smartparking.controller;
 
+import com.smartparking.entity.Client;
+import com.smartparking.entity.Role;
 import com.smartparking.entity.Spot;
 import com.smartparking.model.request.SpotRequest;
 import com.smartparking.model.response.SpotStatisticResponse;
 import com.smartparking.model.response.SpotStatusResponse;
 import com.smartparking.publisher.SpotEventPublisher;
+import com.smartparking.service.ClientService;
 import com.smartparking.service.SpotService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
@@ -18,10 +23,13 @@ import java.util.List;
 public class SpotController {
 
     private final SpotService spotService;
+    @Autowired
+    ClientService clientService;
 
     @Autowired
     SpotEventPublisher spotEventPublisher;
-
+    @Value("${validation.spotnumber.max}")
+    private Long maxSpotNumber;
 
 
     @Autowired
@@ -56,8 +64,6 @@ public class SpotController {
         return spotStatusResponseList;
     }
 
-
-
     @RequestMapping("parkingdetail/{id}/spotstatistic")
     public ResponseEntity<List<SpotStatisticResponse>> getSpotStatistic(
             @PathVariable Long id,
@@ -66,35 +72,60 @@ public class SpotController {
 
 
         List<SpotStatisticResponse> spotStatisticResponseList =
-                spotService.getSpotStatistic(id,Long.parseLong(startTime),Long.parseLong(endTime));
+                spotService.getSpotStatistic(id, Long.parseLong(startTime), Long.parseLong(endTime));
         return new ResponseEntity<>(spotStatisticResponseList, HttpStatus.OK);
     }
 
     @PostMapping("/manager-configuration/spot/save")
     public ResponseEntity<?> save(@RequestBody SpotRequest spotRequest) {
+        Client client = getCurrentUser();
         Spot spot = spotRequest.toSpot();
-        long spotId = 0;
-        if (spot.getId() != null) {
-            spotId = spot.getId();
+        if (isValidNewSpot(spot, client)) {
+            long spotId = 0;
+            if (spot.getId() != null) {
+                spotId = spot.getId();
+            }
+            spotService.save(spot);
+            spotEventPublisher.publishSave(spot, spotId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-        spotService.save(spot);
-        spotEventPublisher.publishSave(spot, spotId);
-        return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @PostMapping("/manager-configuration/spot/delete")
     public ResponseEntity<?> delete(@RequestBody SpotRequest spotRequest) {
+        Client client = getCurrentUser();
         Spot spot = spotRequest.toSpot();
-        spotService.delete(spot);
-        spotEventPublisher.publishDelete(spot);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (client.getProvider().getParkings().stream().filter(parking -> parking.getId().equals(spotRequest.getParkingId())).findFirst().isPresent() ||
+                client.getRole() == Role.SUPERUSER) {
+            spotService.delete(spot);
+            spotEventPublisher.publishDelete(spot);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @GetMapping("/manager-configuration/spotsforparking/{parkingId}")
     public ResponseEntity<List<SpotStatusResponse>> spots(@PathVariable Long parkingId) {
-        return new ResponseEntity<>(spotService.findAllSpotsByParkingIdResponse(parkingId), HttpStatus.OK);
+        Client client = getCurrentUser();
+        if (client.getProvider().getParkings().stream().filter(parking -> parking.getId().equals(parkingId)).findFirst().isPresent() ||
+                client.getRole() == Role.SUPERUSER) {
+            return new ResponseEntity<>(spotService.findAllSpotsByParkingIdResponse(parkingId), HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
+    private Client getCurrentUser() {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        return clientService.findOne(email);
+    }
 
-
+    private Boolean isValidNewSpot(Spot spot, Client client) {
+        return spot.getSpotNumber() != null && spot.getSpotNumber() < maxSpotNumber &&
+                !spot.getParking().getSpots().stream().filter(spot1 -> spot1.getSpotNumber().equals(spot.getSpotNumber())).findFirst().isPresent() &&
+                (client.getProvider().getParkings().contains(spot.getParking()) || client.getRole() == Role.SUPERUSER);
+    }
 }
